refactor: move a /PlanCalculator/PlanPrice folder the PriceWithoutDiscount component
